⚡️ boost levelx parser by polars
