+++ docs/dataset-support/citysim.md
@@ -1,0 +1,7 @@
+# CitySim
+
+## About
+
+## Preparation
+
+The released version of the CitySim dataset is currently hosted on Baidu Cloud, but many folder names are of string format and contain spaces, which can lead to unpredictable issues during access or processing. Therefore, we strongly recommend renaming the dataset files as suggested below before use.
