+++ requirements.txt
@@ -5,6 +5,7 @@ joblib>=1.3.2
 matplotlib>=3.7.0
 numpy>=1.21.5, <=2.2.3
 pandas>=1.5.3
+polars
 protobuf<=3.20.3
 pybind11>=2.10.0
 pygame>=2.5.2
