+++ tactics2d/dataset_parser/__init__.py
@@ -5,7 +5,7 @@
 # @Author: Yueyuan Li
 # @Version: 1.0.0
 
-from .parse_argoverse import ArgoverseParser
+from .parse_argoverse2 import Argoverse2Parser
 from .parse_dlp import DLPParser
 from .parse_interaction import InteractionParser
 from .parse_levelx import LevelXParser
@@ -14,7 +14,7 @@ from .parse_nuplan import NuPlanParser
 from .parse_womd import WOMDParser
 
 __all__ = [
-    "ArgoverseParser",
+    "Argoverse2Parser",
     "DLPParser",
     "InteractionParser",
     "LevelXParser",
