+++ tactics2d/dataset_parser/parse_citysim.py
@@ -1,0 +1,17 @@
+##! python3
+# Copyright (C) 2025, Tactics2D Authors. Released under the GNU GPLv3.
+# @File: parse_citysim.py
+# @Description:
+# @Author: Tactics2D Team
+# @Version:
+
+
+class CitySimParser:
+    """This class implements a parser of the CitySim Dataset.
+
+    !!! quote "Reference"
+    Zheng, Ou, et al. "CitySim: A drone-based vehicle trajectory dataset for safety-oriented research and digital twins." Transportation research record 2678.4 (2024): 606-621.
+    """
+
+    def parse_trajectory(self):
+        return
