--- tactics2d/dataset_parser/womd_proto/pb3/compressed_lidar.proto
@@ -1,37 +1,0 @@
-// @Version: proto3
-// @File: compressed_lidar.proto
-// @Version: 1.0.0
-
-syntax = "proto3";
-
-package waymo.open_dataset;
-
-import "dataset.proto";
-
-message CompressedRangeImage {
-    bytes range_image_delta_compressed = 1;
-    bytes range_image_pose_delta_compressed = 4;
-}
-
-message Metadata {
-    repeated int32 shape = 1;
-    repeated float quant_precision = 2;
-}
-
-message DeltaEncodedData {
-    repeated sint64 residual = 1 [packed = true];
-    repeated uint32 mask = 2 [packed = true];
-    Metadata metadata = 3;
-}
-
-message CompressedLaser {
-    LaserName.Name name = 1;
-    CompressedRangeImage ri_return1 = 2;
-    CompressedRangeImage ri_return2 = 3;
-}
-
-message CompressedFrameLaserData {
-    repeated CompressedLaser lasers = 1;
-    repeated LaserCalibration laser_calibrations = 2;
-    Transform pose = 3;
-}
