--- tactics2d/dataset_parser/womd_proto/pb3/dataset.proto
@@ -1,181 +1,0 @@
-// @Version: proto3
-// @File: dataset.proto
-// @Version: 1.0.0
-
-syntax = "proto3";
-
-package waymo.open_dataset;
-
-import "label.proto";
-import "map.proto";
-import "vector.proto";
-
-message MatrixShape {
-    repeated int32 dims = 1;
-}
-
-message MatrixFloat {
-    repeated float data = 1 [packed = true];
-    MatrixShape shape = 2;
-}
-
-message MatrixInt32 {
-    repeated int32 data = 1 [packed = true];
-    MatrixShape shape = 2;
-}
-
-message CameraName {
-    enum Name {
-        UNKNOWN = 0;
-        FRONT = 1;
-        FRONT_LEFT = 2;
-        FRONT_RIGHT = 3;
-        SIDE_LEFT = 4;
-        SIDE_RIGHT = 5;
-    }
-}
-
-message LaserName {
-    enum Name {
-        UNKNOWN = 0;
-        TOP = 1;
-        FRONT = 2;
-        SIDE_LEFT = 3;
-        SIDE_RIGHT = 4;
-        REAR = 5;
-    }
-}
-
-message Transform {
-    repeated double transform = 1;
-}
-
-message Velocity {
-    float v_x = 1;
-    float v_y = 2;
-    float v_z = 3;
-
-    double w_x = 4;
-    double w_y = 5;
-    double w_z = 6;
-}
-
-message CameraCalibration {
-    CameraName.Name name = 1;
-
-    repeated double intrinsic = 2;
-    Transform extrinsic = 3;
-
-    int32 width = 4;
-    int32 height = 5;
-
-    enum RollingShutterReadOutDirection {
-        UNKNOWN = 0;
-        TOP_TO_BOTTOM = 1;
-        LEFT_TO_RIGHT = 2;
-        BOTTOM_TO_TOP = 3;
-        RIGHT_TO_LEFT = 4;
-        GLOBAL_SHUTTER = 5;
-    }
-
-    RollingShutterReadOutDirection rolling_shutter_direction = 6;
-}
-
-message LaserCalibration {
-    LaserName.Name name = 1;
-
-    repeated double beam_inclinations = 2;
-    double beam_inclination_min = 3;
-    double beam_inclination_max = 4;
-
-    Transform extrinsic = 5;
-}
-
-message Context {
-    string name = 1;
-
-    repeated CameraCalibration camera_calibrations = 2;
-    repeated LaserCalibration laser_calibrations = 3;
-
-    message Stats {
-        message ObjectCount {
-            Label.Type type = 1;
-            int32 count = 2;
-        }
-
-        repeated ObjectCount laser_object_counts = 1;
-        repeated ObjectCount camera_object_counts = 5;
-
-        string time_of_day = 2;
-        string location = 3;
-        string weather = 4;
-    }
-
-    Stats stats = 4;
-}
-
-message RangeImage {
-    bytes range_image_compressed = 2;
-    bytes camera_projection_compressed = 3;
-    bytes range_image_pose_compressed = 4;
-    bytes range_image_flow_compressed = 5;
-    bytes segmentation_label_compressed = 6;
-    MatrixFloat range_image = 1 [deprecated = true];
-}
-
-message CameraSegmentationLabel {
-    int32 panoptic_label_divisor = 1;
-    bytes panoptic_label = 2;
-
-    message InstanceIDToGlobalIDMapping {
-        int32 local_instance_id = 1;
-        int32 global_instance_id = 2;
-        bool is_tracked = 3;
-    }
-
-    repeated InstanceIDToGlobalIDMapping instance_id_to_global_id_mapping = 3;
-
-    string sequence_id = 4;
-
-    bytes num_cameras_covered = 5;
-}
-
-message CameraImage {
-    CameraName.Name name = 1;
-    bytes image = 2;
-    Transform pose = 3;
-    Velocity velocity = 4;
-    double pose_timestamp = 5;
-    double shutter = 6;
-    double camera_trigger_time = 7;
-    double camera_readout_done_time = 8;
-    CameraSegmentationLabel camera_segmentation_label = 10;
-}
-
-message CameraLabels {
-    CameraName.Name name = 1;
-    repeated Label labels = 2;
-}
-
-message Laser {
-    LaserName.Name name = 1;
-    RangeImage ri_return1 = 2;
-    RangeImage ri_return2 = 3;
-}
-
-message Frame {
-    Context context = 1;
-    int64 timestamp_micros = 2;
-    Transform pose = 3;
-
-    repeated CameraImage images = 4;
-    repeated Laser lasers = 5;
-    repeated Label laser_labels = 6;
-    repeated CameraLabels projected_lidar_labels = 9;
-    repeated CameraLabels camera_labels = 8;
-
-    repeated Polygon2dProto no_label_zones = 7;
-
-    repeated MapFeature map_features = 10;
-    Vector3d map_pose_offset = 11;
-}
