--- tactics2d/dataset_parser/womd_proto/pb3/keypoint.proto
@@ -1,62 +1,0 @@
-// @Version: proto3
-// @File: keypoint.proto
-// @Version: 1.0.0
-
-syntax = "proto3";
-
-package waymo.open_dataset.keypoints;
-
-import "vector.proto";
-
-message KeypointVisibility {
-    bool is_occluded = 1;
-}
-
-message Keypoint2d {
-    Vector2d location_px = 1;
-    KeypointVisibility visibility = 2;
-
-}
-
-message Keypoint3d {
-    Vector3d location_m = 1;
-    KeypointVisibility visibility = 2;
-}
-
-enum KeypointType {
-    KEYPOINT_TYPE_UNSPECIFIED = 0;
-    KEYPOINT_TYPE_NOSE = 1;
-    KEYPOINT_TYPE_LEFT_SHOULDER = 5;
-    KEYPOINT_TYPE_LEFT_ELBOW = 6;
-    KEYPOINT_TYPE_LEFT_WRIST = 7;
-    KEYPOINT_TYPE_LEFT_HIP = 8;
-    KEYPOINT_TYPE_LEFT_KNEE = 9;
-    KEYPOINT_TYPE_LEFT_ANKLE = 10;
-    KEYPOINT_TYPE_RIGHT_SHOULDER = 13;
-    KEYPOINT_TYPE_RIGHT_ELBOW = 14;
-    KEYPOINT_TYPE_RIGHT_WRIST = 15;
-    KEYPOINT_TYPE_RIGHT_HIP = 16;
-    KEYPOINT_TYPE_RIGHT_KNEE = 17;
-    KEYPOINT_TYPE_RIGHT_ANKLE = 18;
-    KEYPOINT_TYPE_FOREHEAD = 19;
-    KEYPOINT_TYPE_HEAD_CENTER = 20;
-  }
-
-  message CameraKeypoint {
-    KeypointType type = 1;
-    Keypoint2d keypoint_2d = 2;
-    Keypoint3d keypoint_3d = 3;
-  }
-
-message CameraKeypoints {
-    repeated CameraKeypoint keypoint = 1;
-}
-
-message LaserKeypoint {
-    KeypointType type = 1;
-    Keypoint3d keypoint_3d = 2;
-}
-
-message LaserKeypoints {
-    repeated LaserKeypoint keypoint = 1;
-}
