--- tactics2d/dataset_parser/womd_proto/pb3/label.proto
@@ -1,89 +1,0 @@
-// @Version: proto3
-// @File: label.proto
-// @Version: 1.0.0
-
-syntax = "proto3";
-
-package waymo.open_dataset;
-
-import "keypoint.proto";
-
-message Label {
-    message Box {
-        double center_x = 1;
-        double center_y = 2;
-        double center_z = 3;
-
-        double length = 5;
-        double width = 4;
-        double height = 6;
-
-        double heading = 7;
-
-        enum Type {
-            TYPE_UNKNOWN = 0;
-            TYPE_3D = 1;
-            TYPE_2D = 2;
-            TYPE_AA_2D = 3;
-        }
-    }
-
-    Box box = 1;
-
-    message Metadata {
-        double speed_x = 1;
-        double speed_y = 2;
-        double speed_z = 5;
-        double accel_x = 3;
-        double accel_y = 4;
-        double accel_z = 6;
-    }
-
-    Metadata metadata = 2;
-
-    enum Type {
-        TYPE_UNKNOWN = 0;
-        TYPE_VEHICLE = 1;
-        TYPE_PEDESTRIAN = 2;
-        TYPE_SIGN = 3;
-        TYPE_CYCLIST = 4;
-    }
-
-    Type type = 3;
-
-    string id = 4;
-
-    enum DifficultyLevel {
-        UNKNOWN = 0;
-        LEVEL_1 = 1;
-        LEVEL_2 = 2;
-    }
-
-    DifficultyLevel detection_difficulty_level = 5;
-    DifficultyLevel tracking_difficulty_level = 6;
-
-    int32 num_lidar_points_in_box = 7;
-    int32 num_top_lidar_points_in_box = 13;
-
-    oneof keypoints_oneof {
-        keypoints.LaserKeypoints laser_keypoints = 8;
-        keypoints.CameraKeypoints camera_keypoints = 9;
-    }
-
-    message Association {
-        string laser_object_id = 1;
-    }
-
-    Association association = 10;
-
-    string most_visible_camera_name = 11;
-
-    Box camera_synced_box = 12;
-}
-
-message Polygon2dProto {
-    repeated double x = 1;
-    repeated double y = 2;
-
-    string id = 3;
-}
