--- tactics2d/dataset_parser/womd_proto/pb3/vector.proto
@@ -1,18 +1,0 @@
-// @Version: proto3
-// @File: vector.proto
-// @Version: 1.0.0
-
-syntax = "proto3";
-
-package waymo.open_dataset;
-
-message Vector2d {
-    double x = 1;
-    double y = 2;
-}
-
-message Vector3d {
-    double x = 1;
-    double y = 2;
-    double z = 3;
-}
